# Conflicts:
#	Engine/src/manager/WorldManager.java
#	UI/src/display/Console.java
#	UI/src/manager/EngineAgent.java

+++ Engine/src/simulation/properties/property/api/PropertyType.java
@@ -1,5 +1,6 @@
 package simulation.properties.property.api;
 
 public enum PropertyType {
-    INT, DOUBLE, BOOLEAN, STRING
+    DECIMAL, FLOAT, BOOLEAN, STRING
 }
+
